[DSIP-54][Master] Use ClusterManager to manage the cluster in master (#16263)

+++ dolphinscheduler-common/src/main/java/org/apache/dolphinscheduler/common/model/WorkerHeartBeat.java
@@ -29,6 +29,6 @@ import lombok.experimental.SuperBuilder;
 public class WorkerHeartBeat extends BaseHeartBeat implements HeartBeat {
 
     private int workerHostWeight; // worker host weight
-    private int threadPoolUsage; // worker waiting task count
+    private double threadPoolUsage; // worker waiting task count
 
 }
